requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
h5py>=3.0.0
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cu118
torchvision>=0.15.0 --index-url https://download.pytorch.org/whl/cu118
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0
biopython>=1.79
mlflow>=2.0.0
peft>=0.4.0
gputil
transformers
goatools
adjustText
onnx
onnxscript
